cloud_functions/requirements.txt
click==8.1.3
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.3.2
functions-framework==3.5.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-events==0.10.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==21.3
pathtools==0.1.2
pyparsing==3.1.0
setuptools==68.0.0
watchdog==1.0.2
Werkzeug==2.3.6